website/api/requirements.txt
click==8.1.3
colorama==0.4.6
flash==1.0.3
Flask==2.2.3
flask-login==0.6.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
pytest==7.2.2
python-dotenv==0.21.1
PyYAML==6.0
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.13.0
